src/backend/pyproject.toml
[project]
name = "agora-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic-settings>=2.2",
  "requests>=2.31",
  "weaviate-client>=4.20.1",
  "openai>=1.30",
  "tiktoken>=0.7",
  "dspy-ai>=2.4.9",
]

[tool.ruff]
line-length = 100
target-version = "py310"
